# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Inventario de productos",
    "version": "1.0",
    "title": "INVENTARIO",
    "contact": {}
  },
  "host": "localhost:5000",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/productos": {
      "get": {
        "summary": "Con esto listas todos tus productos",
        "tags": [
          "LISTAR O AÑADIR PRODUCTOS"
        ],
        "operationId": "/productos",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "post": {
        "summary": "Con esto ingresas nuevos productos",
        "tags": [
          "LISTAR O AÑADIR PRODUCTOS"
        ],
        "operationId": "Post/productos",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/~1productosrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/productos/{id}": {
      "delete": {
        "summary": "Con esto eliminas solo un producto",
        "tags": [
          "ELIMINAR ACTULIZAR O BUSCAR"
        ],
        "operationId": "/productos/<int:id>",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id del producto",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "summary": "Con esto actualizas el estado de un solo producto",
        "tags": [
          "ELIMINAR ACTULIZAR O BUSCAR"
        ],
        "operationId": "Put/productos/<int:id>",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id del producto",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/~1productos~1%3Cint%3Aid%3Erequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "summary": "Con esto consultas el estado de un solo producto",
        "tags": [
          "ELIMINAR ACTULIZAR O BUSCAR"
        ],
        "operationId": "/productos/<int:id>1",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id del producto",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    }
  },
  "definitions": {
    "/productosrequest": {
      "title": "/productosrequest",
      "example": {
        "nombre": "Leche de almendras",
        "precio": 3.5,
        "cantidad": 1
      },
      "type": "object",
      "properties": {
        "nombre": {
          "type": "string"
        },
        "precio": {
          "type": "number",
          "format": "double"
        },
        "cantidad": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "nombre",
        "precio",
        "cantidad"
      ]
    },
    "/productos/<int:id>request": {
      "title": "/productos/<int:id>request",
      "example": {
        "nombre": "1 KG de Azucar",
        "precio": 4,
        "cantidad": 1
      },
      "type": "object",
      "properties": {
        "nombre": {
          "type": "string"
        },
        "precio": {
          "type": "number",
          "format": "double"
        },
        "cantidad": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "nombre",
        "precio",
        "cantidad"
      ]
    }
  },
  "tags": [
    {
      "name": "LISTAR O AÑADIR PRODUCTOS"
    },
    {
      "name": "ELIMINAR ACTULIZAR O BUSCAR"
    }
  ]
}